@@ -3,25 +3,26 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
     "@material-ui/core": "^3.9.4",
     "@material-ui/lab": "^3.0.0-alpha.30",
-    "@tanem/svg-injector": "^8.0.36",
+    "@tanem/svg-injector": "^8.0.63",
     "@types/react-syntax-highlighter": "^11.0.4",
     "googleapis": "^39.2.0",
-    "mdi-material-ui": "^5.21.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "mdi-material-ui": "^5.29.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-markdown": "^4.3.1",
-    "react-router-dom": "^5.1.2",
+    "react-responsive": "^8.1.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.0.0",
     "react-scroll": "^1.8.1",
     "react-svg": "^9.0.4",
     "react-syntax-highlighter": "^12.2.1",
-    "react-transition-group": "^4.3.0",
+    "react-transition-group": "^4.4.1",
     "styled-components": "^5.1.1",
     "typeface-roboto": "0.0.54",
-    "typescript": "^3.7.4",
-    "@babel/plugin-proposal-optional-chaining": "^7.11.0"
+    "typescript": "^3.9.7"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,10 +48,12 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^2.1.1"
+    "gh-pages": "^2.2.0"
   },
   "homepage": "http://atude.github.io/atude-ghp",
   "babel": {
-    "plugins": ["@babel/plugin-proposal-optional-chaining"]
+    "plugins": [
+      "@babel/plugin-proposal-optional-chaining"
+    ]
   }
 }
